+++ CHANGELOG.md
@@ -7,9 +7,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.0] - 2018-12-04
+
 ### Added
 
 - Additional Checks: All relevant statements are present
+- Additional Checks: All relevant statements are in correct order
+- Additional Checks: Person or entity statements should be used in an ownership-or-control statement
+- Additional Checks: scheme should be an org-id code (This requires lib-cove 0.2.0 or higher)
+- Statistic: % of Interests of each type
+
+### Changed
+
+- Upgrades lib-cove needed to 0.3.1
 
 ## [0.1.0] - 2018-11-22